+++ docs/changelog.rst
@@ -1,6 +1,15 @@
 Change Log
 ==========
 
+v0.4 (not yet released)
+-----------------------
+
+* Constructors: ``create_from_clauses``, ``create_characteristic``
+  `#41 <https://github.com/msoeken/kitty/pull/41>`_
+
+* CNF generation: ``cnf_characteristic``
+  `#41 <https://github.com/msoeken/kitty/pull/41>`_
+
 v0.3 (February 25, 2018)
 ------------------------